Rewrite CLI to use a stateful default profile.
